refactor: api 모듈명 변경 및 중복된 import 제거
